requirements.txt
# ============================================
# RAG-SQL - Dependencias
# ============================================

# ----- LangChain Core -----
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# ----- Base de datos -----
psycopg2-binary>=2.9.0

# ----- Configuración -----
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ----- API REST -----
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# ----- Tokens y conteo -----
tiktoken>=0.5.0

# ----- Cache -----
redis>=5.0.0

# ----- Semantic Cache (Qdrant) -----
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# ----- Testing -----
pytest>=7.0.0
pytest-asyncio>=0.21.0

# ----- Utilidades -----
aiohttp>=3.8.0

# ============================================
# Proveedores LLM Opcionales
# (instalar según el proveedor que uses)
# ============================================
# pip install langchain-anthropic    # Para Claude
# pip install langchain-groq         # Para Groq
# pip install langchain-google-genai # Para Gemini
# pip install langchain-ollama       # Para Ollama local
